WIP: tests/inst: Add destructive test framework
This adds infrastructure to the Rust test suite for destructive
tests, and adds a new `transactionality` test which runs
rpm-ostree in a loop (along with `ostree-finalize-staged`) and
repeatedly kills them.

The main goal here is to flush out any "logic errors".  I plan
to further extend this to reboots and then force poweroffs.